Use a local function for 'webkitMatchesSelector'
That's preferable to polluting the jQuery namespace, as per
https://github.com/atom/space-pen/pull/29#issuecomment-38200024

There might be a cleaner way (or a better namespace?) to stick
that particular function in - I'm no expert on the space-pen
codebase, feel free to refactor further / put it in the right
place. The functionality is equivalent to the first #29 commit.