Replace attached/removed hooks with versions based on custom elements
The hooks are now called ::attached and ::detached instead of
::afterAttach and ::beforeRemove for reasons described below.

This eliminates the need to use SpacePen’s modified version of jQuery to
attach and detach SpacePen views in order to get the lifecycle hooks.
Since we are using direct DOM manipulation in Atom core this is
essential to making SpacePen views compatible going forward.

The semantics are different and more limited. The custom element
lifecycle callbacks only allow detection of attachment and detachment
to/from the actual DOM, not also another detached element. We also
can’t distinguish between jQuery’s concept of “detachment” (meaning the
element will be reattached) and “removal” (meaning the element is gone
forever). Instead we offer only detachment and attachment detection.