Warn about atom.commands when emitting/subscribing to event with a `:`
This adds a warning to event subscription via jQuery if the event name
contains a `:`, because there’s a good chance it is an attempt to
subscribe to a command. Since there are legitimate uses of a `:` that
don’t involve commands, I’ve chosen just to warn and not to fail.

I’ve also updated `$::trigger` with the same approach for symmetry.